Update all templates layout include register apps with create, update delete pages stopped unlogged in user to cerate or update and delete

+++ cambr_proj/post/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 
@@ -23,36 +23,15 @@ def list_post(request):
         posts = paginator.page(paginator.num_pages)
     context = {
         'query_set': posts,
-        # 'title': 'title',
-        # 'content': 'content',
-        # 'create_date': 'Public Date',
     }
 
     return render(request, 'posts.html', context)
 
 
-def backend_posts(request):
-    posts_list = Post.objects.all().order_by('-create_date')
-    paginator = Paginator(posts_list, 7)  # Show 25 posts per page
-    page = request.GET.get('page')
-    try:
-        posts = paginator.page(page)
-    except PageNotAnInteger:
-        # If page is not an integer, deliver first page.
-        posts = paginator.page(1)
-    except EmptyPage:
-        # If page is out of range (e.g. 9999), deliver last page of results.
-        posts = paginator.page(paginator.num_pages)
-    context = {
-        'query_set': posts,
-        # 'title': 'title',
-        # 'content': 'content',
-        # 'create_date': 'Public Date',
-    }
-    return render(request, 'backend_post_list.html', context)
-
-
 def create_post(request):
+    # This will stop user to acces to create page without login
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         obj_form = form.save(commit=False)
@@ -61,12 +40,7 @@ def create_post(request):
         # message successfully created
         messages.success(request, 'Successfully created ')
         return HttpResponseRedirect(obj_form.get_absolute_url())
-    # else:
-    # 	messages.error(request, 'Error the post not created ')
-    # if request.method == 'POST':
-    # 	title = request.POST.get('title')
-    # 	content = request.POST.get('content')
-    # 	print request.POST
+
     context = {
         'form': form,
     }
@@ -82,11 +56,14 @@ def post_details(request, slug=None):
         'content': instance_obj.content,
         'create_date': instance_obj.create_date,
         'image': instance_obj.image,
+        'slug': instance_obj.slug,
     }
     return render(request, 'post_detail.html', context)
 
 
 def update_post(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
 
     form = PostForm(request.POST or None,
@@ -98,18 +75,41 @@ def update_post(request, slug=None):
         # message successfully updated
         messages.success(request, 'Successfully Updated ')
         return HttpResponseRedirect(obj_form.get_absolute_url())
-    # else:
-    # 	messages.error(request, 'Error not updated ')
+
     context = {
         'title': instance.title,
-        'instance': instance,
+        'instance': instance.content,
         'form': form,
+        'image': instance.image,
+        'slug': instance.slug,
     }
-    return render(request, 'create_post.html', context)
+    return render(request, 'post_update.html', context)
 
 
 def delete_post(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, 'Successfully Deleted ')
-    return redirect('posts:list_post')
+    return redirect('post:list_post')
+
+# def backend_posts(request):
+#     posts_list = Post.objects.all().order_by('-create_date')
+#     paginator = Paginator(posts_list, 7)  # Show 25 posts per page
+#     page = request.GET.get('page')
+#     try:
+#         posts = paginator.page(page)
+#     except PageNotAnInteger:
+#         # If page is not an integer, deliver first page.
+#         posts = paginator.page(1)
+#     except EmptyPage:
+#         # If page is out of range (e.g. 9999), deliver last page of results.
+#         posts = paginator.page(paginator.num_pages)
+#     context = {
+#         'query_set': posts,
+#         # 'title': 'title',
+#         # 'content': 'content',
+#         # 'create_date': 'Public Date',
+#     }
+#     return render(request, 'backend_post_list.html', context)
